IntervalCollection: Remove interval helper interface
