+++ packages/dds/sequence/src/intervalIndex/endpointInRangeIndex.ts
@@ -8,11 +8,10 @@
 import { Client, PropertyAction, RedBlackTree } from "@fluidframework/merge-tree/internal";
 
 import {
-	IIntervalHelpers,
 	ISerializableInterval,
 	IntervalType,
 	SequenceInterval,
-	sequenceIntervalHelpers,
+	createSequenceInterval,
 } from "../intervals/index.js";
 import { ISharedString } from "../sharedString.js";
 
@@ -29,25 +28,20 @@ import {
  * Provide additional APIs to support efficiently querying a collection of intervals whose endpoints fall within a specified range.
  * @internal
  */
-export interface IEndpointInRangeIndex<TInterval extends ISerializableInterval>
-	extends IntervalIndex<TInterval> {
+export interface IEndpointInRangeIndex<SequenceInterval extends ISerializableInterval>
+	extends IntervalIndex<SequenceInterval> {
 	/**
 	 * @returns an array of all intervals contained in this collection whose endpoints locate in the range [start, end] (includes both ends)
 	 */
-	findIntervalsWithEndpointInRange(start: number, end: number): TInterval[];
+	findIntervalsWithEndpointInRange(start: number, end: number): SequenceInterval[];
 }
 
-export class EndpointInRangeIndex<TInterval extends ISerializableInterval>
-	implements IEndpointInRangeIndex<TInterval>
-{
+export class EndpointInRangeIndex implements IEndpointInRangeIndex<SequenceInterval> {
 	private readonly intervalTree;
 
-	constructor(
-		private readonly client: Client,
-		private readonly helpers: IIntervalHelpers<TInterval>,
-	) {
-		this.intervalTree = new RedBlackTree<TInterval, TInterval>(
-			(a: TInterval, b: TInterval) => {
+	constructor(private readonly client: Client) {
+		this.intervalTree = new RedBlackTree<SequenceInterval, SequenceInterval>(
+			(a: SequenceInterval, b: SequenceInterval) => {
 				const compareEndsResult = a.compareEnd(b);
 				if (compareEndsResult !== 0) {
 					return compareEndsResult;
@@ -71,25 +65,25 @@ export class EndpointInRangeIndex<TInterval extends ISerializableInterval>
 		);
 	}
 
-	public add(interval: TInterval): void {
+	public add(interval: SequenceInterval): void {
 		this.intervalTree.put(interval, interval);
 	}
 
-	public remove(interval: TInterval): void {
+	public remove(interval: SequenceInterval): void {
 		this.intervalTree.remove(interval);
 	}
 
-	public findIntervalsWithEndpointInRange(start: number, end: number): TInterval[] {
+	public findIntervalsWithEndpointInRange(start: number, end: number): SequenceInterval[] {
 		if (start <= 0 || start > end || this.intervalTree.isEmpty()) {
 			return [];
 		}
-		const results: TInterval[] = [];
-		const action: PropertyAction<TInterval, TInterval> = (node) => {
+		const results: SequenceInterval[] = [];
+		const action: PropertyAction<SequenceInterval, SequenceInterval> = (node) => {
 			results.push(node.data);
 			return true;
 		};
 
-		const transientStartInterval = this.helpers.create(
+		const transientStartInterval = createSequenceInterval(
 			"transient",
 			start,
 			start,
@@ -97,7 +91,7 @@ export class EndpointInRangeIndex<TInterval extends ISerializableInterval>
 			IntervalType.Transient,
 		);
 
-		const transientEndInterval = this.helpers.create(
+		const transientEndInterval = createSequenceInterval(
 			"transient",
 			end,
 			end,
@@ -121,5 +115,5 @@ export function createEndpointInRangeIndex(
 	sharedString: ISharedString,
 ): IEndpointInRangeIndex<SequenceInterval> {
 	const client = (sharedString as unknown as { client: Client }).client;
-	return new EndpointInRangeIndex<SequenceInterval>(client, sequenceIntervalHelpers);
+	return new EndpointInRangeIndex(client);
 }
