+++ packages/dds/sequence/src/intervalIndex/endpointIndex.ts
@@ -7,12 +7,7 @@
 
 import { Client, RedBlackTree } from "@fluidframework/merge-tree/internal";
 
-import {
-	IIntervalHelpers,
-	IntervalType,
-	SequenceInterval,
-	sequenceIntervalHelpers,
-} from "../intervals/index.js";
+import { createSequenceInterval, IntervalType, SequenceInterval } from "../intervals/index.js";
 import { ISharedString } from "../sharedString.js";
 
 import { type SequenceIntervalIndex } from "./intervalIndex.js";
@@ -37,17 +32,14 @@ export interface IEndpointIndex extends SequenceIntervalIndex {
 export class EndpointIndex implements IEndpointIndex {
 	private readonly endIntervalTree: RedBlackTree<SequenceInterval, SequenceInterval>;
 
-	constructor(
-		private readonly client: Client,
-		private readonly helpers: IIntervalHelpers,
-	) {
+	constructor(private readonly client: Client) {
 		this.endIntervalTree = new RedBlackTree<SequenceInterval, SequenceInterval>((a, b) =>
 			a.compareEnd(b),
 		);
 	}
 
 	public previousInterval(pos: number): SequenceInterval | undefined {
-		const transientInterval = this.helpers.create(
+		const transientInterval = createSequenceInterval(
 			"transient",
 			pos,
 			pos,
@@ -61,7 +53,7 @@ export class EndpointIndex implements IEndpointIndex {
 	}
 
 	public nextInterval(pos: number): SequenceInterval | undefined {
-		const transientInterval = this.helpers.create(
+		const transientInterval = createSequenceInterval(
 			"transient",
 			pos,
 			pos,
@@ -88,5 +80,5 @@ export class EndpointIndex implements IEndpointIndex {
  */
 export function createEndpointIndex(sharedString: ISharedString): IEndpointIndex {
 	const client = (sharedString as unknown as { client: Client }).client;
-	return new EndpointIndex(client, sequenceIntervalHelpers);
+	return new EndpointIndex(client);
 }
