+++ packages/dds/sequence/src/intervalIndex/overlappingIntervalsIndex.ts
@@ -13,11 +13,10 @@ import {
 
 import { IntervalNode, IntervalTree } from "../intervalTree.js";
 import {
-	IIntervalHelpers,
 	ISerializableInterval,
 	IntervalType,
 	SequenceInterval,
-	sequenceIntervalHelpers,
+	createSequenceInterval,
 } from "../intervals/index.js";
 import { ISharedString } from "../sharedString.js";
 
@@ -74,11 +73,9 @@ export interface ISequenceOverlappingIntervalsIndex extends SequenceIntervalInde
 export class OverlappingIntervalsIndex implements ISequenceOverlappingIntervalsIndex {
 	protected readonly intervalTree = new IntervalTree<SequenceInterval>();
 	protected readonly client: Client;
-	protected readonly helpers: IIntervalHelpers;
 
-	constructor(client: Client, helpers: IIntervalHelpers) {
+	constructor(client: Client) {
 		this.client = client;
-		this.helpers = helpers;
 	}
 
 	public map(fn: (interval: SequenceInterval) => void) {
@@ -111,7 +108,7 @@ export class OverlappingIntervalsIndex implements ISequenceOverlappingIntervalsI
 				});
 			}
 		} else {
-			const transientInterval: SequenceInterval = this.helpers.create(
+			const transientInterval: SequenceInterval = createSequenceInterval(
 				"transient",
 				start ?? "start",
 				end ?? "end",
@@ -187,7 +184,7 @@ export class OverlappingIntervalsIndex implements ISequenceOverlappingIntervalsI
 		) {
 			return [];
 		}
-		const transientInterval = this.helpers.create(
+		const transientInterval = createSequenceInterval(
 			"transient",
 			start,
 			end,
@@ -216,5 +213,5 @@ export function createOverlappingIntervalsIndex(
 	sharedString: ISharedString,
 ): ISequenceOverlappingIntervalsIndex {
 	const client = (sharedString as unknown as { client: Client }).client;
-	return new OverlappingIntervalsIndex(client, sequenceIntervalHelpers);
+	return new OverlappingIntervalsIndex(client);
 }
