+++ packages/dds/sequence/src/intervals/index.ts
@@ -10,7 +10,6 @@ export {
 	IntervalOpType,
 	IntervalType,
 	IntervalDeltaOpType,
-	IIntervalHelpers,
 	IntervalStickiness,
 	ISerializableIntervalPrivate,
 	SerializedIntervalDelta,
@@ -18,11 +17,9 @@ export {
 	endReferenceSlidingPreference,
 	startReferenceSlidingPreference,
 } from "./intervalUtils.js";
-export { Interval, intervalHelpers } from "./interval.js";
 export {
 	SequenceInterval,
 	SequenceIntervalClass,
 	createSequenceInterval,
 	createPositionReferenceFromSegoff,
-	sequenceIntervalHelpers,
 } from "./sequenceInterval.js";
