+++ packages/dds/sequence/src/intervals/intervalUtils.ts
@@ -7,7 +7,6 @@
 
 import { ISequencedDocumentMessage } from "@fluidframework/driver-definitions/internal";
 import {
-	Client,
 	PropertiesManager,
 	PropertySet,
 	SlidingPreference,
@@ -214,38 +213,6 @@ export type CompressedSerializedInterval =
 			IntervalStickiness,
 	  ]
 	| [number | "start" | "end", number | "start" | "end", number, IntervalType, PropertySet];
-
-/**
- * @sealed
- * @internal
- */
-export interface IIntervalHelpers<TInterval extends ISerializableInterval = SequenceInterval> {
-	/**
-	 *
-	 * @param label - label of the interval collection this interval is being added to. This parameter is
-	 * irrelevant for transient intervals.
-	 * @param start - numerical start position of the interval
-	 * @param end - numerical end position of the interval
-	 * @param client - client creating the interval
-	 * @param intervalType - Type of interval to create. Default is SlideOnRemove
-	 * @param op - If this create came from a remote client, op that created it. Default is undefined (i.e. local)
-	 * @param fromSnapshot - If this create came from loading a snapshot. Default is false.
-	 * @param startSide - The side on which the start position lays. See
-	 * {@link @fluidframework/merge-tree#SequencePlace} for additional context
-	 * @param endSide - The side on which the end position lays. See
-	 * {@link @fluidframework/merge-tree#SequencePlace} for additional context
-	 */
-	create(
-		label: string,
-		start: SequencePlace | undefined,
-		end: SequencePlace | undefined,
-		client: Client | undefined,
-		intervalType: IntervalType,
-		op?: ISequencedDocumentMessage,
-		fromSnapshot?: boolean,
-		useNewSlidingBehavior?: boolean,
-	): TInterval;
-}
 
 /**
  * Determines how an interval should expand when segments are inserted adjacent
